Wandering Trades player heads now load last; ConsoleUtils new method logMessage; ResourceManager now doesn't overwrites files that already exist.

+++ modules/craftera_suite-core/src/main/java/com/github/scorchedpsyche/craftera_suite/modules/main/ResourcesManager.java
@@ -50,7 +50,15 @@ public class ResourcesManager
                 File destinationFile = new File(destinationPath);
 
                 try {
-                    FileUtils.copyInputStreamToFile(resourceToBeCopied, destinationFile);
+                    if( !destinationFile.exists() )
+                    {
+                        FileUtils.copyInputStreamToFile(resourceToBeCopied, destinationFile);
+                        consoleUtils.logMessage(
+                                "File copied: " + ChatColor.YELLOW + destinationFile);
+                    } else {
+                        consoleUtils.logMessage(
+                                "File exists and won't be copied: " + ChatColor.YELLOW + destinationFile);
+                    }
                 } catch (IOException e) {
                     // Failed to create file
                     consoleUtils.logError(
